+++ packages/codemode/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @opengeni/codemode
 
+## 0.4.3
+
+### Patch Changes
+
+- Updated dependencies [083387e]
+- Updated dependencies [11913b7]
+  - @opengeni/contracts@1.0.0
+
 ## 0.4.2
 
 ### Patch Changes