+++ packages/documents/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @opengeni/documents
 
+## 0.5.40
+
+### Patch Changes
+
+- Updated dependencies [083387e]
+- Updated dependencies [11913b7]
+  - @opengeni/contracts@1.0.0
+  - @opengeni/db@1.0.0
+  - @opengeni/config@0.16.3
+  - @opengeni/storage@0.2.95
+
 ## 0.5.39
 
 ### Patch Changes